# Fast-track `@threads` when `nthreads() == 1`