@yuyichao 2019-05-29
This check should not happen at macro expansion time.