@raminammour 2019-05-29
Hello,

The only way I have found to work around #15276 is to make sure that the code with `@threads, nthreads()=1` runs almost as fast as the non-threaded code. I even suspect that is the reason for the slowdown (`@code_warntype` shows a `Core.Box`) in your example.

Would this not make it harder to detect that the closure bug is preventing speedup in multi-threaded code?

Cheers!