@staticfloat 2019-05-29
@raminammour while #15276 is something that can get triggered by code like this (and is in this case), this is fixing something independent of that.  Yes, the `Box` slows this down, and when I change this PR according to the feedback above we continue to trigger the `Box` problems, and thereby don't get the same speedup; but we do get a decent amount of speedup still, because we are eliminating a different source of slowdown.

If you want to test for #15276 style problems, I suggest you do a more direct test than relying on the fact that `@threads` introduces a closure that induces boxing; that may not be true forever (indeed I hope it's not!).

I've updated the original PR message with new performance metrics and a slightly updated benchmark script.