@raminammour 2019-05-30
Just FYI, the benchmarks on my system are different:

```

using BenchmarkTools, Base.Threads

function func(val)
    local sum = 0*(1 .^ val)
    for idx in 1:100
        sum += idx.^val
    end
    return sum
end

function func_threaded_let(val)
    local sum = 0*(1 .^ val)
    @threads for idx in 1:100
        let sum=sum
        sum += idx.^val
        end
    end
    return sum
end
function func_threaded(val)
    local sum = 0*(1 .^ val)
    @threads for idx in 1:100
        sum += idx.^val
    end
    return sum
end

@show @benchmark func(2.0)
@show @benchmark func_threaded(2.0)
@show @benchmark func_threaded_let(2.0);
versioninfo()

#= In[134]:30 =# @benchmark(func(2.0)) = Trial(2.333 μs)
#= In[134]:31 =# @benchmark(func_threaded(2.0)) = Trial(4.551 μs)
#= In[134]:32 =# @benchmark(func_threaded_let(2.0)) = Trial(2.568 μs)
Julia Version 1.0.2
Commit d789231e99 (2018-11-08 20:11 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin14.5.0)
  CPU: Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.0 (ORCJIT, haswell)
```
Hope this helps.