@jebej 2019-05-30
It seems most of the slowdown is due to the closure bug, I get:
```
#= REPL[11]:1 =# @benchmark(func(2.0)) = Trial(129.937 ns)

#= REPL[12]:1 =# @benchmark(func_threaded(2.0)) = Trial(3.716 μs)

#= REPL[13]:1 =# @benchmark(func_threaded_let(2.0)) = Trial(442.848 ns)
```
On `v1.1.1`.