@staticfloat 2019-05-30
The `let` version that @raminammour posted does not calculate the same thing; it never stores the `sum` value:

```
julia> func(2.0)
338350.0

julia> func_threaded(2.0)
338350.0

julia> func_threaded_let(2.0)
0.0
```

If you look at the `@code_native` of `func_threaded_let()` versus `func_threaded()`, you will see that the `_let()` variant contains less than half as much code; the optimizer is able to get rid of a lot of work because you're not using the output in any way by creating a new binding within the `@threads` block.