@raminammour 2019-05-30
Sorry, my bad, here is a `Ref` version that does the right thing, and is fast as it avoids #15276. (I suspect the time would have been much faster if the whole calculation in the `let` version was avoided). I may still be confused though :)

```using BenchmarkTools, Base.Threads

function func(val)
     sum = 0*(1 .^ val)
    for idx in 1:100
        sum += idx.^val
    end
    return sum
end

function func_threaded_ref(val)
     sum = Ref(0*(1 .^ val))
    @threads for idx in 1:100
        sum[] += idx.^val
    end
    return sum[]
end
function func_threaded(val)
     sum = 0*(1 .^ val)
    @threads for idx in 1:100
        sum += idx.^val
    end
    return sum
end

@btime func(2.0)
@btime func_threaded(2.0)
@btime func_threaded_ref(2.0);
versioninfo()

  2.520 μs (0 allocations: 0 bytes)
  4.623 μs (203 allocations: 3.20 KiB)
  2.440 μs (2 allocations: 64 bytes)
Julia Version 1.0.2
Commit d789231e99 (2018-11-08 20:11 UTC)
Platform Info:
  OS: macOS (x86_64-apple-darwin14.5.0)
  CPU: Intel(R) Core(TM) i7-4980HQ CPU @ 2.80GHz
  WORD_SIZE: 64
  LIBM: libopenlibm
  LLVM: libLLVM-6.0.0 (ORCJIT, haswell)
```