@staticfloat 2019-05-30
Yes, that version does indeed remove much of the leftover performance gap; I'll include that in my benchmark above, but notice that you're still getting MUCH slower timings for even the serial case because you're using Julia 1.0; use Julia 1.1 or even better `master` and you'll see a large difference in performance.  Your `Ref` change doesn't quite eliminate the speed difference, although it does get it closer:

Before this PR:
```
#= /Users/sabae/src/julia/speedtest.jl:27 =# @benchmark(func(2.0)) = Trial(81.144 ns)
#= /Users/sabae/src/julia/speedtest.jl:28 =# @benchmark(func_threaded(2.0)) = Trial(8.739 μs)
#= /Users/sabae/src/julia/speedtest.jl:29 =# @benchmark(func_threaded_ref(2.0)) = Trial(6.890 μs)
```

After this PR:
```
#= /Users/sabae/src/julia/speedtest.jl:27 =# @benchmark(func(2.0)) = Trial(81.176 ns)
#= /Users/sabae/src/julia/speedtest.jl:28 =# @benchmark(func_threaded(2.0)) = Trial(3.775 μs)
#= /Users/sabae/src/julia/speedtest.jl:29 =# @benchmark(func_threaded_ref(2.0)) = Trial(1.674 μs)
```