@staticfloat 2019-06-03
@JeffBezanson, @yuyichao any further comments?  I'm unsure why `@threads` with a single thread retains such a slowdown (I assume because of inference goblins) but the speedup here is not insignificant on its own.