@staticfloat 2019-06-19
Pinging @JeffBezanson and @yuyichao again to see if there are any further comments, if not I think we should merge this, as it's a straight performance win when using `@threads` with only one thread.