@yuyichao 2019-06-19
How much speed up you get if you use https://github.com/JuliaLang/julia/pull/21452/files#diff-7198cded2577e0bdeb563f0f2713347bR69 instead?

Also, that branch is somehow changed to use the latest world in https://github.com/JuliaLang/julia/pull/30838, which is presumably to be consistent with the threading branch. If that's the case and is intended, then this should do that as well which means that it cannot be compiled / inlined ahead of time.

The threading branch is changed to use the latest world in https://github.com/JuliaLang/julia/pull/31398/files#diff-5a6699a5aa7cf07be50461e3c7f68262L693 and in particular the code was added in https://github.com/JuliaLang/julia/commit/d9d8d4ca6e607550fbf914c4940b66707ffca0dd. Why is that? That seems to be fairly different from the semantics before and the semantics of a normal for loop. Is that needed for something in particular? Otherwise, I don't really see the point of making this change.

Another reason calling the function is preferred is to make sure the semantics of the loop is actually the same. This means that the single threaded `@thread` loop should have the same scope rules (in a function) and the same limitations as a normal one. This way you can actually test it with a single thread and be fairly sure that the code could run with multiple threads (short of other true thread related bugs).

Also, is the nested thread loop hack still needed? Isn't the point of `partr` to get rid of it? (That was what I had in mind when adding that code anyway...)