@vtjnash 2019-06-19
Yeah, the `jl_threading_run` function now is just a statically compiled copy of some Julia code. That's pretty slow (and awkward / hard to maintain), so we should move that function into Julia