@vchuravy 2019-06-19
I agree that we shouldn't performance hack the `tid=1` case, I have found it pretty important to debug issues early on by being able to say: Is this threaded loop slow on a single thread? What happens if I remove it.